requirements.txt
# requirements
ccxt>=4.3.0
pandas>=2.0.0
numpy>=1.25.0
ta>=0.10.2
python-dotenv>=1.0.0
dash>=2.17.0
plotly>=5.20.0
requests>=2.31.0
scipy>=1.10.0